Add map

+++ srcs/bresenham.c
@@ -1,16 +1,16 @@
 #include <wolf3d.h>
 
-void    plotLineLow(t_win *win, int p0[2], int p1[2])
+void    plotLineLow(t_win *win, double p0[2], double p1[2])
 {
-  int dx = p1[0] - p0[0];
-  int dy = p1[1] - p0[1];
-  int yi = 1;
+  double dx = p1[0] - p0[0];
+  double dy = p1[1] - p0[1];
+  double yi = 1;
   if (dy < 0)
   {
     yi = -1;
     dy = -dy;
     }
-  int D = 2 * dy - dx;
+  double D = 2 * dy - dx;
 
   while (p0[0] <= p1[0])
   {
@@ -25,17 +25,17 @@ void    plotLineLow(t_win *win, int p0[2], int p1[2])
   }
 }
 
-void    plotLineHigh(t_win *win, int p0[2], int p1[2])
+void    plotLineHigh(t_win *win, double p0[2], double p1[2])
 {
-  int dx = p1[0] - p0[0];
-  int dy = p1[1] - p0[1];
-  int xi = 1;
+  double dx = p1[0] - p0[0];
+  double dy = p1[1] - p0[1];
+  double xi = 1;
   if (dx < 0)
   {
     xi = -1;
     dx = -dx;
   }
-  int D = 2 * dx - dy;
+  double D = 2 * dx - dy;
 
   while (p0[1] <= p1[1])
   {
@@ -50,7 +50,7 @@ void    plotLineHigh(t_win *win, int p0[2], int p1[2])
   }
 }
 
-void    bresenham(t_win *win, int p0[2], int p1[2])
+void    bresenham(t_win *win, double p0[2], double p1[2])
 {
       if (ABS((p1[1] - p0[1])) < ABS((p1[0] - p0[0])))
     {
